✓ adds cast
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1606 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ WORK_01/codecnv/eucucs2.c
@@ -39,7 +39,7 @@ UINT codecnv_euctoucs2(UINT16 *lpOutput, UINT cchOutput, const char *lpInput, UI
 	else
 	{
 		// String mode
-		nLength = euctoucs2(lpOutput, cchOutput - 1, lpInput, strlen(lpInput));
+		nLength = euctoucs2(lpOutput, cchOutput - 1, lpInput, (UINT)strlen(lpInput));
 		if (lpOutput)
 		{
 			lpOutput[nLength] = '\0';
